Switch to finding DCL class and add settings name to BuildConfig
