--- hiddenapi/stubs/src/main/java/android/ext/settings/app/AswRestrictMemoryDynCodeLoading.java
@@ -1,4 +1,0 @@
-package android.ext.settings.app;
-
-public class AswRestrictMemoryDynCodeLoading {
-}
