+++ zygisk/src/main/kotlin/org/matrix/vector/core/Main.kt
@@ -1,12 +1,7 @@
 package org.matrix.vector.core
 
-import android.content.pm.ApplicationInfo
-import android.ext.settings.app.AswRestrictMemoryDynCodeLoading
 import android.os.IBinder
 import android.os.Process
-import de.robv.android.xposed.XC_MethodReplacement
-import de.robv.android.xposed.XposedBridge
-import de.robv.android.xposed.XposedHelpers
 import org.lsposed.lspd.service.ILSPApplicationService
 import org.lsposed.lspd.util.Utils
 import org.matrix.vector.BuildConfig
@@ -40,43 +35,8 @@ object Main {
             ParasiticManagerSystemHooker.start()
         }
 
-        if (niceName == "system" || niceName == "com.android.settings") {
-            try {
-                // Force GrapheneOS to allow changing the restriction on Dynamic Code Loading, and
-                // force-disable it for the settings app and the shell
-                XposedBridge.hookAllMethods(
-                    AswRestrictMemoryDynCodeLoading::class.java,
-                    "getImmutableValue",
-                    object : XC_MethodReplacement() {
-                        override fun replaceHookedMethod(param: MethodHookParam<*>?): Any? {
-                            val appInfo = param?.args[2] as? ApplicationInfo?
-
-                            // Settings has to always be allowed so that it can be patched as well.
-                            if (appInfo != null && listOf(
-                                    "com.android.settings",
-                                    "com.android.shell"
-                                ).contains(appInfo.packageName)
-                            ) {
-                                return false
-                            }
-
-                            // All system apps are configurable (null)
-                            if (appInfo != null && (appInfo.flags and ApplicationInfo.FLAG_SYSTEM) != 0) {
-                                return null
-                            }
-
-                            // Defer to original implementation for other apps
-                            return XposedBridge.invokeOriginalMethod(
-                                param?.method, param?.thisObject, param?.args
-                            )
-                        }
-                    }
-                )
-            } catch (e: XposedHelpers.ClassNotFoundError) {
-                // Ignore, assuming we are not on GrapheneOS
-            } catch (e: Exception) {
-                Utils.logE("Unknown error patching Graphene", e)
-            }
+        if (isSystem || niceName == BuildConfig.GrapheneSettingsPackageName) {
+            ParasiticManagerHooker.patchGrapheneDCLRestriction()
         }
 
         // Initialize Xposed bridge components
